Solve Makefile issue

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: kpiacent <kpiacent@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/02/25 19:46:03 by kpiacent          #+#    #+#             */
-/*   Updated: 2016/02/26 09:39:07 by kpiacent         ###   ########.fr       */
+/*   Updated: 2016/04/13 19:07:33 by kpiacent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	int		i;
+	size_t		i;
 
 	i = 0;
 	while (i < n)
